chore: marketplace html refactor
