+++ src/components/Marketplace/Marketplace.jsx
@@ -3,9 +3,14 @@ import { getCollectionsByChain } from "./collections";
 import NFTCollectionItems from "./NFTCollectionItems";
 import { paddingLRContent, } from "../../GlobalStyles";
 import { MainChainID } from "../../MglNftMetadata";
+import { descriptionStyle } from "../../GlobalStyles";
+import { chainIdToNameAndLogo } from "../Chains/Chains";
 
 function Marketplace() {
   const NFTCollections = getCollectionsByChain(MainChainID);
+  const marketPlaceChainId = MainChainID;
+  const chainName = chainIdToNameAndLogo.get(marketPlaceChainId)[0];
+  const chainLogo = chainIdToNameAndLogo.get(marketPlaceChainId)[1];
 
   return (
     <div style={{
@@ -28,6 +33,18 @@ function Marketplace() {
         })
         }
         <br />
+      </div>
+      <div style={{
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "center",
+        ...descriptionStyle,
+      }}>
+        NFTs collection on&nbsp;
+        <span style={{
+        }}>{chainName}</span>
+        &nbsp;
+        {chainLogo}
       </div>
     </div>
   );
